Erfan/delete entities (#180)

+++ src/Datiss.Budget.Services/IncomeCurrentCofficientService.cs
@@ -212,13 +212,20 @@ namespace Datiss.Budget.Services
                                    .Where(x => x.OrganizationId == organizationId)
                                    .ToListAsync();
 
-            var childrens = await getChildren(organizationId, yearId);
+            IEnumerable<IncomeCurrentCofficient> childrens = new IncomeCurrentCofficient[] { };
+
+            if (organization.Type == OrganizationType.County || organization.Type == OrganizationType.Root)
+            {
+                childrens = await getChildren(organizationId, yearId);
+            }
 
             if (self.Count() == 0 && childrens.Count() == 0)
                 throw new DeleteNullRecordException();
 
             _dbSet.RemoveRange(self);
-            _dbSet.RemoveRange(childrens);
+
+            if (childrens.Any())
+                _dbSet.RemoveRange(childrens);
 
             var result = new OrganizationDeleteDataResult
             {
